Game Completed
This is probably everything that we need it goes from start to finish if start from the main screen. I will post a build soon so we can test for bugs.

+++ OverCootsed/Assets/Scripts/Ingredients/IngredientStats.cs
@@ -1,9 +1,10 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using UnityEngine.UI;
 
 public class IngredientStats : MonoBehaviour
 {
-    public int totalGrams;
-    public int strength;
+    public float totalGrams;
+    public float strength;
 }
